feat(doctors): update doctor cms, controller, and requests

+++ app/Http/Controllers/Backsite/DoctorController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 
 // use library here
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 // request
@@ -15,6 +17,7 @@ use App\Http\Requests\Doctor\UpdateDoctorRequest;
 // use everything here
 use Gate;
 use Auth;
+use File;
 
 // use model here
 use App\Models\Operational\Doctor;
@@ -41,6 +44,8 @@ class DoctorController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('doctor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // for table grid
         $doctor = Doctor::orderBy('created_at', 'desc')->get();
 
@@ -71,6 +76,23 @@ class DoctorController extends Controller
         // get all request from frontsite
         $data = $request->all();
 
+        // upload process here
+        // do not forget php artisan storage:link to create shortcut so blade can access the photo
+        $path = public_path('app/public/assets/file-doctor');
+        if (!File::isDirectory($path)) {
+            $response = Storage::makeDirectory('public/assets/file-doctor');
+        }
+
+        // change file locations
+        if (isset($data['photo'])) {
+            $data['photo'] = $request->file('photo')->store(
+                'assets/file-doctor',
+                'public'
+            );
+        } else {
+            $data['photo'] = "";
+        }
+
         // store to database
         $doctor = Doctor::create($data);
 
@@ -99,7 +121,7 @@ class DoctorController extends Controller
      */
     public function edit(Doctor $doctor)
     {
-        abort_if(Gate::denies('doctor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('doctor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         // for select2 = ascending a to z
         $specialist = Specialist::orderBy('name', 'asc')->get();
@@ -119,6 +141,32 @@ class DoctorController extends Controller
         // get all request from frontsite
         $data = $request->all();
 
+        // re format before push to table
+        $data['fee'] = str_replace(',', '', $data['fee']);
+        $data['fee'] = str_replace('IDR ', '', $data['fee']);
+
+        // upload process here
+        // change format photo
+        if (isset($data['photo'])) {
+
+            // first checking old photo to delete from storage
+            $get_item = $doctor['photo'];
+
+            // change file locations
+            $data['photo'] = $request->file('photo')->store(
+                'assets/file-doctor',
+                'public'
+            );
+
+            // delete old photo from storage
+            $data_old = 'storage/' . $get_item;
+            if (File::exists($data_old)) {
+                File::delete($data_old);
+            } else {
+                File::delete('storage/app/public/' . $get_item);
+            }
+        }
+
         // update to database
         $doctor->update($data);
 
@@ -134,6 +182,18 @@ class DoctorController extends Controller
      */
     public function destroy(Doctor $doctor)
     {
+        abort_if(Gate::denies('doctor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        // first checking old file to delete from storage
+        $get_item = $doctor['photo'];
+
+        $data = 'storage/' . $get_item;
+        if (File::exists($data)) {
+            File::delete($data);
+        } else {
+            File::delete('storage/app/public/' . $get_item);
+        }
+
         $doctor->forceDelete();
 
         alert()->success('Success Message', 'Successfully deleted doctor');
